refactor: rename network status module
