+++ feeluown/gui/components/__init__.py
@@ -13,4 +13,4 @@ from .song_tag import SongSourceTag  # noqa
 from .collections import CollectionListView  # noqa
 from .player_progress import PlayerProgressSliderAndLabel  # noqa
 from .recommendation_panel import Panel, RecPlaylistsPanel  # noqa
-from .proxy_status import NetworkStatusButton  # noqa
+from .network_status import NetworkStatusButton  # noqa
